funcion create user with email and password funciona
